Split main.c for better readibiliy
