+++ src/main.h
@@ -1,5 +1,10 @@
 #pragma once
 
+#include "crypto.h"
+#include "vetClauseUstream.h"
+#include "vetUstream.h"
+#include "vetDisplay.h"
+
 typedef struct internalStorage_t {
     uint8_t dataAllowed;
     uint8_t multiClauseAllowed;
@@ -10,15 +15,3 @@ extern const internalStorage_t N_storage_real;
 #define N_storage (*(volatile internalStorage_t *) PIC(&N_storage_real))
 
 extern void ui_idle(void);
-
-extern volatile char fullAddress[43];
-extern volatile char fullAmount[50];
-extern volatile char maxFee[60];
-extern volatile bool dataPresent;
-extern volatile bool multipleClauses;
-
-unsigned int io_seproxyhal_touch_settings();
-unsigned int io_seproxyhal_touch_exit();
-unsigned int io_seproxyhal_touch_tx_ok();
-unsigned int io_seproxyhal_touch_address_ok();
-unsigned int io_seproxyhal_touch_cancel();
