+++ src/ui/ui_callback.h
@@ -1,0 +1,6 @@
+#pragma once
+
+unsigned int io_seproxyhal_touch_exit();
+unsigned int io_seproxyhal_touch_tx_ok();
+unsigned int io_seproxyhal_touch_address_ok();
+unsigned int io_seproxyhal_touch_cancel();
